Minor refactor requests and response models
