--- PiggyBank/Data/AccountResponse.swift
@@ -1,13 +1,0 @@
-import Foundation
-
-final class AccountResponse: Codable {
-    
-    let id: Int
-    let type: Int
-    let title: String
-    let currency: String
-    let balance: Double
-    let isArchived: Bool
-    let isDeleted: Bool
-    
-}
