--- PiggyBank/Data/CategoryResponse.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-public final class CategoryResponse: Codable {
-    
-    let id: Int
-    let title: String
-    let hexColor: String
-    let type: Int
-    let isArchived: Bool
-    let isDeleted: Bool
-    
-}
