--- PiggyBank/Data/CreateUpdateCategoryRequest.swift
@@ -1,17 +1,0 @@
-import Foundation
-
-final class CreateUpdateCategoryRequest: Codable {
-    
-    let title: String
-    let hexColor: String
-    let type: Int
-    let isArchived: Bool
-    
-    init(title: String, hexColor: String, type: Int, isArchived: Bool) {
-        self.title = title
-        self.hexColor = hexColor
-        self.type = type
-        self.isArchived = isArchived
-    }
-    
-}
