--- PiggyBank/Data/CurrencyResponse.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-final class CurrencyResponse: Codable {
-    
-    let code: String
-    let symbol: String
-    
-}
