--- PiggyBank/Data/ReportsByCategoryRequest.swift
@@ -1,10 +1,0 @@
-//
-//  ReportsByCategoryRequest.swift
-//  PiggyBank
-//
-
-public struct ReportsByCategoryRequest: Codable {
-    let type: Int
-    let from: String
-    let to: String
-}
