--- PiggyBank/Data/ReportsByCategoryResponse.swift
@@ -1,11 +1,0 @@
-//
-//  ReportsByCategoryResponse.swift
-//  PiggyBank
-//
-
-public struct ReportsByCategoryResponse: Codable {
-    let categoryId: Int
-    let categoryTitle: String
-    let categoryHexColor: String
-    let amount: Int
-}
