--- PiggyBank/Data/UserCredentialsResponse.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-final class UserCredentialsResponse: Decodable {
-    
-    let accessToken: String
-}
