+++ PiggyBank/Network/Models/Category.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+enum Category {
+
+    final class Response: Codable {
+        let id: Int
+        let title: String
+        let hexColor: String
+        let type: Int
+        let isArchived: Bool
+        let isDeleted: Bool
+    }
+
+}
