+++ PiggyBank/Network/Models/Currency.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum Currency {
+
+    final class Response: Codable {
+        let code: String
+        let symbol: String
+    }
+
+}
