+++ PiggyBank/Network/Models/SignIn.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+enum SignIn {
+    struct Request: Codable {
+        let userName: String
+        let password: String
+
+        init(nickname: String, password: String) {
+            self.userName = nickname
+            self.password = password
+        }
+    }
+}
