+++ PiggyBank/Network/Models/UserCredentials.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum UserCredentials {
+
+    struct Response: Decodable {
+        let accessToken: String
+    }
+}
